Incorporating code review comments
