+++ java-example-extension/run.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: MIT-0
 
 EXTENSION_NAME=$1
 LAMBDA_FUNCTION=$2
